pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = "Analysis of the integration of AdePT in Gaussino simulations."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc[ssh]>=3.63.0",
    "marimo>=0.14.16",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
